+++ examples/aspnetcore8-mvc/Models/ErrorViewModel.cs
@@ -1,0 +1,9 @@
+namespace aspnetcore8_mvc.Models
+{
+  public class ErrorViewModel
+  {
+    public string? RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+  }
+}