+++ src/Bugsnag.ConfigurationSection/Configuration.cs
@@ -428,7 +428,7 @@ namespace Bugsnag.ConfigurationSection
       {
         try
         {
-          if (Polyfills.String.IsNullOrWhiteSpace(ProxyAddress)) return null;
+          if (String.IsNullOrWhiteSpace(ProxyAddress)) return null;
 
           // we should probably store this so we don't try to create a new one each time this is accessed
           if (!string.IsNullOrEmpty(ProxyUsername) && !string.IsNullOrEmpty(ProxyPassword))