+++ src/Bugsnag/Properties/AssemblyInfo.cs
@@ -1,6 +1,3 @@
 using System;
-#if NET45
-using System.Security;
-#endif
 
 [assembly: CLSCompliant(true)]