+++ tests/Bugsnag.AspNet.WebApi.Tests/WebHostTests.cs
@@ -2,6 +2,7 @@ using Bugsnag.Tests;
 using System;
 using System.Linq;
 using System.Net.Http;
+using System.Threading.Tasks;
 using System.Web.Http;
 using Xunit;
 
@@ -20,7 +21,7 @@ namespace Bugsnag.AspNet.WebApi.Tests
     }
 
     [Fact]
-    public async void Test()
+    public async Task Test()
     {
       var bugsnagServer = new TestServer();
 
@@ -40,7 +41,7 @@ namespace Bugsnag.AspNet.WebApi.Tests
       var responses = await bugsnagServer.Requests(1);
 
       Assert.Single(responses);
-      Assert.Contains("Bugsnag is great!", responses.Single());
+      Assert.Contains("Bugsnag is great!", responses.Single().Body);
     }
   }
 }