+++ tests/Bugsnag.Tests/SerializerTests.cs
@@ -58,6 +58,20 @@ namespace Bugsnag.Tests
       Assert.NotNull(json);
     }
 
+    [Fact]
+    public void CanSerializeException()
+    {
+      try
+      {
+        throw new System.Exception("Serialize me");
+      }
+      catch (System.Exception exception)
+      {
+        var json = Serializer.SerializeObject(exception);
+        Assert.NotNull(json);
+      }
+    }
+
     private class Circular
     {
       public string Name { get; set; }
@@ -87,6 +101,7 @@ namespace Bugsnag.Tests
       yield return new object[] { new FilterableTestObject { { "password", "password" } }, null, 0 };
       yield return new object[] { new FilterableTestObject { { "username", "password" } }, new string[] { "password" }, 0 };
       yield return new object[] { new FilterableTestObject { { "password", "password" }, { "credit_card_number", "number" } }, new string[] { "password", "credit_card_number" }, 2 };
+      yield return new object[] { new FilterableTestObject { { "Password", "password" }, { "Credit_Card_Number", "number" } }, new string[] { "password", "credit_card_number" }, 2 };
     }
   }
 }